made the requirement of the actual password for the possibilities to change both password and e-mails as well as putting the common parts of the change in them in a single method.
Created and re-organized the tests in Users.

Signed-off-by Bruno de Amorim Campos<amorim.software@gmail.com>

made the requirement of the actual password for the possibilities to change both password and e-mails as well as putting the common parts of the change in them in a single method.

Created and re-organized the tests in Users.

Signed-off-by Bruno de Amorim Campos<amorim.software@gmail.com>